Scattering

The electric field measured at sensor positioned at $\mathbf{R} = R\hat{n}$ results from the sum of all dipoles located at $\mathbf{r}_j$ and their respective states $\beta_j$. The exact formula depends on the model and regime.

Electric Field

Scalar

regime = :near_field $ E{sc}(\mathbf{R}, t) = +i\frac{\Gamma}{2}\sumj \frac{e^{ ik0|\mathbf{R} - \mathbf{r}j| }}{k0|\mathbf{R} - \mathbf{r}j|}\beta_j(t) $

regime = :far_field $ E{sc}(\mathbf{R}, t) \approx +i\frac{\Gamma}{2} \frac{e^{ ik0R }}{k0R}\sumj exp( -ik0|\hat{n} \cdot \mathbf{r}j| )\beta_j(t) $

Vectorial

regime = :near_field $ E{vec}(\mathbf{R}, t) = -i\frac{\Gamma}{2}\sumj\sum{\eta}G{\mu,\eta}(\mathbf{R}-\mathbf{r}j)\betaj^{\eta}(t) $

regime = :far_field $ E^\mu{vec}(\mathbf{R},t) \approx -i\frac{\Gamma}{2}\cdot\frac{3}{2} \frac{e^{ik0R}}{k0R}\sumj\sum\eta(\delta{\mu, \eta} - \hat{n}\mu\hat{n}\eta^*)exp(-ik0\hat{\mathbf{n}}\cdot\mathbf{r}j)\beta_j^\eta(t) $

Mean Field

regime = :near_field|:far_field $ \mathbf{E}{mf} = \mathbf{E}{sc} $

Intensity

Scalar

regime = :near_field|:far_field $ I{sc}(\vec{R},t) = |\mathbf{E}{sc}|^2 $

Vectorial

regime = :near_field|:far_field $ I{vec}(\hat{n},t) = |\mathbf{E}{vec}|^2 = \sum\mu|E^\mu{vec}|^2 $

Mean Field

regime = :near_field $ I{mf}(\mathbf{R},t) = I{sc}(\mathbf{R}, t) +\frac{\Gamma^2}{(2k0)^2} \left [ \sum{j} \frac{- |\betaj|^2 + \frac{1+\langle \sigmaj^z \rangle }{2}}{|R-rj|^2} \right ]. $ `regime = :farfield` $ I{mf}(\mathbf{R},t) = I{sc}(\mathbf{R}, t) + \frac{\Gamma^2}{(2k0R)^2}\sum{j=1}^N \left ( -|\betaj|^2 + \frac{1 + \langle \sigmaj^z \rangle }{2}\right ) $

Functions

CoupledDipoles.scattered_electric_field — Function

scattered_electric_field(problem, atomic_states, sensor_positions; regime=:far_field)

Returns a Matrix{ComplexF64} with value of the Eletric Laser + Electric Scattered from atoms

problem: LinearOptics or NonLinearOptics

atomic_states: β for Scalar/Vectorial Model, or [β,z] for Mean Field Model

sensor_positions: matrix with measurement points

Note:

Scalar problem returns a Matrix and not a Vector, to maintain consistency with Vectorial problem that necessary returns a Matrix, where each column has the [Ex, Ey, Ez] components of the field.

Also, even for single sensor, returns a Matrix of one element.

Example

using CoupledDipoles, Random
Random.seed!(111)
N = 5
kR, kh = 1.0, 1.0
atoms = Atom(Cylinder(), N, kR, kh)

s, Δ = 1e-5, 1.0
laser = Laser(PlaneWave3D(), s, Δ; polarization=[1,0,0])

problem_scalar = LinearOptics(Scalar(), atoms, laser)
problem_vectorial = LinearOptics(Vectorial(), atoms, laser)

atomic_states_scalar = steady_state(problem_scalar)
atomic_states_vectorial = steady_state(problem_vectorial)

## 1 sensor
Random.seed!(222)
nSensors = 1
sensor = Matrix(rand(3, nSensors)) # '3' == sensor in position in 3D space
scattered_electric_field(problem_scalar, atomic_states_scalar, sensor)
scattered_electric_field(problem_vectorial, atomic_states_vectorial, sensor)

## 10 sensors
Random.seed!(333)
nSensors = 10
sensor = rand(3, nSensors) # '3' == sensor in position in 3D space
scattered_electric_field(problem_scalar, atomic_states_scalar, sensor)
scattered_electric_field(problem_vectorial, atomic_states_vectorial, sensor)

CoupledDipoles.laser_and_scattered_intensity — Function

laser_and_scattered_intensity(problem, atomic_states, sensor_positions; regime=:far_field)

Returns a Vector{Float64} with value of the |Electric Laser + Electric Scattered|^2 from atoms

problem: LinearOptics or NonLinearOptics

atomic_states: β for Scalar/Vectorial Model, or [β,z] for Mean Field Model

sensor_positions: matrix with measurement points

CoupledDipoles.scattered_intensity — Function

scattered_intensity(problem, atomic_states, sensor_positions; regime=:far_field)

Returns a Vector{Float64} with value of the |Electric Scattered|^2 from atoms

problem: LinearOptics or NonLinearOptics

atomic_states: β for Scalar/Vectorial Model, or [β,z] for Mean Field Model

sensor_positions: matrix with measurement points

Example

using CoupledDipoles, Random
Random.seed!(111)
N = 5
kR, kh = 1.0, 1.0
atoms = Atom(Cylinder(), N, kR, kh)

s, Δ = 1e-5, 1.0
laser = Laser(PlaneWave3D(), s, Δ; polarization=[1,0,0])

problem_scalar = LinearOptics(Scalar(), atoms, laser)
problem_vectorial = LinearOptics(Vectorial(), atoms, laser)

atomic_states_scalar = steady_state(problem_scalar)
atomic_states_vectorial = steady_state(problem_vectorial)

## 1 sensor
Random.seed!(222)
nSensors = 1
sensor = Matrix(rand(3, nSensors)) # '3' == sensor in position in 3D space
scattered_intensity(problem_scalar, atomic_states_scalar, sensor)
scattered_intensity(problem_vectorial, atomic_states_vectorial, sensor)

## 10 sensors
Random.seed!(333)
nSensors = 10
sensor = rand(3, nSensors) # '3' == sensor in position in 3D space
scattered_intensity(problem_scalar, atomic_states_scalar, sensor)
scattered_intensity(problem_vectorial, atomic_states_vectorial, sensor)

CoupledDipoles.get_intensity_over_an_angle — Function

get_intensity_over_an_angle(problem::LinearOptics{Scalar}, atoms_states::AbstractVector, θ_range::AbstractVector)

Used for the same state and many angles (suitable for cbs)

using CoupledDipoles, Random
Random.seed!(111)
N = 5
kR, kh = 1.0, 1.0
atoms = Atom(Cylinder(), N, kR, kh)

s, Δ = 1e-5, 1.0
laser = Laser(PlaneWave3D(), s, Δ; polarization=[1,0,0])

problem_scalar = LinearOptics(Scalar(), atoms, laser)
atomic_states_scalar = steady_state(problem_scalar)

θ_range = deg2rad.(range(0, 360, length=30))
get_intensity_over_an_angle(problem_scalar, atomic_states_scalar, θ_range)

get_intensity_over_an_angle(problem::LinearOptics{Scalar}, atoms_states::Vector, θ::Number; tol=exp10(-7.4))

Used for the single angle and single single state (most probably user case).

Example:

using CoupledDipoles, Random
Random.seed!(111)
N = 5
kR, kh = 1.0, 1.0
atoms = Atom(Cylinder(), N, kR, kh)

s, Δ = 1e-5, 1.0
laser = Laser(PlaneWave3D(), s, Δ; polarization=[1,0,0])

problem_scalar = LinearOptics(Scalar(), atoms, laser)
atomic_states_scalar = steady_state(problem_scalar)

θ = deg2rad(48)
get_intensity_over_an_angle(problem_scalar, atomic_states_scalar, θ)

get_intensity_over_an_angle(problem::LinearOptics{Scalar}, atoms_states::Vector{Vector{ComplexF64}}, θ::Number; tol=exp10(-7.4), exact_solution=false)

Used for the same angle and different states (for example, the output of time_evolution).

Example:

using CoupledDipoles, Random
Random.seed!(111)
N = 5
kR, kh = 1.0, 1.0
atoms = Atom(Cylinder(), N, kR, kh)

s, Δ = 1e-5, 1.0
laser = Laser(PlaneWave3D(), s, Δ; polarization=[1,0,0])

problem_scalar = LinearOptics(Scalar(), atoms, laser)
u0 = default_initial_condition(problem_scalar)
tspan = (0.0, 10.0)
solutions = time_evolution(problem_scalar, u0, tspan)
states = solutions.u

θ = deg2rad(48)
get_intensity_over_an_angle(problem_scalar, states, θ)