Steady State

LinearOptics has a well defined analytical solution through a linear system of equations. In Scalar case, one get an Array solution, and in Vectorial case, one gets a Matrix, where each collum represents the x,y,z-components.

using CoupledDipoles
r =[1 2 0;
    1 0 1.0]
atoms = Atom(Cube(), Array(transpose(r)), 10)
laser = Laser(PlaneWave3D(), 1e-6, 1.0)
problem = LinearOptics(Scalar(), atoms, laser)
βₛₛ = steady_state(problem) # N-array

2-element Vector{ComplexF64}:
  0.00037248446780549374 + 0.000622054602877758im
 -0.00018637781550957864 + 0.0004955719149269015im

# (...) same as Scalar case
problem = LinearOptics(Vectorial(), atoms, laser)
βₛₛ = steady_state(problem) # 3xN-matrix

3×2 Matrix{ComplexF64}:
 -0.000347004+0.000192822im  -0.000293954-5.47096e-5im
          0.0-0.0im                   0.0-0.0im
          0.0-0.0im                   0.0-0.0im

NonLinearOptics has not well defined solution, therefore steady_state apply time_evolution function over the period tspan = (0, 500) and return the final state.

The solution is an Array of size 2N, corresponding to [$\langle \sigma^- \rangle$ $\langle \sigma^z \rangle$].

# (...) same as Scalar case
problem = NonLinearOptics(MeanField(), atoms, laser)
βₛₛ = steady_state(problem) # 2N-array

4-element Vector{ComplexF64}:
   0.0003724840316892595 + 0.0006220539481834492im
 -0.00018637785676937116 + 0.0004955716963157544im
      -0.999998948607958 - 1.1267637399346156e-16im
     -0.9999994393436622 - 9.367153808429973e-17im

CoupledDipoles.steady_state — Function

steady_state(problem::LinearOptics{Scalar})

Solve x=G\Ω, with default interaction_matrix and laser_field.

steady_state(problem::LinearOptics{Vectorial})

Solve x=-G\Ω, with default interaction_matrix and laser_field. The solution x is reshaped as a 3xN matrix.

steady_state(problem::NonLinearOptics{MeanField}; tmax = 250.0, reltol = 1e-11, abstol = 1e-10, m = 90, ode_solver = false)