Steady State

LinearOptics has a well defined analytical solution through a linear system of equations. In Scalar case, one get an Array solution, and in Vectorial case, one gets a Matrix, where each collum represents the x,y,z-components.

using CoupledDipoles
r =[1 2 0;
    1 0 1.0]
atoms = Atom(Cube(), Array(transpose(r)), 10)
laser = Laser(PlaneWave3D(), 1e-6, 1.0)
problem = LinearOptics(Scalar(), atoms, laser)
βₛₛ = steady_state(problem) # N-array

problem = LinearOptics(Vectorial(), atoms, laser)
βₛₛ = steady_state(problem) # 3xN-matrix

NonLinearOptics has not well defined solution, therefore steady_state apply time_evolution function over the period tspan = (0, 500) and return the final state.

The solution is an Array of size 2N, corresponding to [$\langle \sigma^- \rangle$ $\langle \sigma^z \rangle$].

using CoupledDipoles
r =[1 2 0;
    1 0 1.0]
atoms = Atom(Cube(), Array(transpose(r)), 10)
laser = Laser(PlaneWave3D(), 1e-6, 1.0)
problem = NonLinearOptics(MeanField(), atoms, laser)
βₛₛ = steady_state(problem) # 2N-array

CoupledDipoles.steady_state — Function

steady_state(problem::LinearOptics{Scalar})

Solve x=G\Ω, with default interaction_matrix and laser_field.

steady_state(problem::LinearOptics{Vectorial})

Solve x=-G\Ω, with default interaction_matrix and laser_field. The solution x is reshaped as a 3xN matrix.

steady_state(problem::NonLinearOptics{MeanField}; tmax = 250.0, reltol = 1e-10, abstol = 1e-10, m = 90, ode_solver = false)